fix: move parse_installable into impl Source
